add the ticks in the top-part of-calendar.
