+++ src/components/calendar/calendar.tsx
@@ -10,7 +10,7 @@ import {
 } from "../../helpers/date-helper";
 import { DateSetup } from "../../types/date-setup";
 import styles from "./calendar.module.css";
-import { addToDate } from "../../helpers/date-helper";
+// import { addToDate } from "../../helpers/date-helper";
 
 export type CalendarProps = {
   dateSetup: DateSetup;
@@ -80,6 +80,13 @@ export const Calendar: React.FC<CalendarProps> = ({
             y2Line={headerHeight}
             xText={xText}
             yText={topDefaultHeight * 0.9}
+            tasks={tasks}
+            rowHeight={rowHeight}
+            svgWidth={svgWidth}
+            dates={dates}
+            todayColor={todayColor}
+            columnWidth={columnWidth}
+            rtl={rtl}
           />
         );
       }
@@ -87,41 +94,41 @@ export const Calendar: React.FC<CalendarProps> = ({
     return [topValues, bottomValues];
   };
 
-  let y = 0;
-  const gridRows: ReactChild[] = [];
-  const rowLines: ReactChild[] = [
-    <line
-      key="RowLineFirst"
-      x="0"
-      y1={0}
-      x2={svgWidth}
-      y2={0}
-      className={styles.gridRowLine}
-    />,
-  ];
-  for (const task of tasks) {
-    gridRows.push(
-      <rect
-        key={"Row" + task.id}
-        x="0"
-        y={y}
-        width={svgWidth}
-        height={rowHeight}
-        className={styles.gridRow}
-      />
-    );
-    rowLines.push(
-      <line
-        key={"RowLine" + task.id}
-        x="0"
-        y1={y + rowHeight}
-        x2={svgWidth}
-        y2={y + rowHeight}
-        className={styles.gridRowLine}
-      />
-    );
-    y += rowHeight;
-  }
+  // let y = 0;
+  // const gridRows: ReactChild[] = [];
+  // const rowLines: ReactChild[] = [
+  //   <line
+  //     key="RowLineFirst"
+  //     x="0"
+  //     y1={0}
+  //     x2={svgWidth}
+  //     y2={0}
+  //     className={styles.gridRowLine}
+  //   />,
+  // ];
+  // for (const task of tasks) {
+  //   gridRows.push(
+  //     <rect
+  //       key={"Row" + task.id}
+  //       x="0"
+  //       y={y}
+  //       width={svgWidth}
+  //       height={rowHeight}
+  //       className={styles.gridRow}
+  //     />
+  //   );
+  //   rowLines.push(
+  //     <line
+  //       key={"RowLine" + task.id}
+  //       x="0"
+  //       y1={y + rowHeight}
+  //       x2={svgWidth}
+  //       y2={y + rowHeight}
+  //       className={styles.gridRowLine}
+  //     />
+  //   );
+  //   y += rowHeight;
+  // }
 
   const getCalendarValuesForMonth = () => {
     const topValues: ReactChild[] = [];
@@ -160,6 +167,13 @@ export const Calendar: React.FC<CalendarProps> = ({
             y2Line={topDefaultHeight}
             xText={xText}
             yText={topDefaultHeight * 0.9}
+            tasks={tasks}
+            rowHeight={rowHeight}
+            svgWidth={svgWidth}
+            dates={dates}
+            todayColor={todayColor}
+            columnWidth={columnWidth}
+            rtl={rtl}
           />
         );
       }
@@ -206,6 +220,13 @@ export const Calendar: React.FC<CalendarProps> = ({
               y2Line={topDefaultHeight}
               xText={columnWidth * i + columnWidth * weeksCount * 0.5}
               yText={topDefaultHeight * 0.9}
+              tasks={tasks}
+              rowHeight={rowHeight}
+              svgWidth={svgWidth}
+              dates={dates}
+              todayColor={todayColor}
+              columnWidth={columnWidth}
+              rtl={rtl}
             />
           );
         }
@@ -257,6 +278,13 @@ export const Calendar: React.FC<CalendarProps> = ({
                 0.5
             }
             yText={topDefaultHeight * 0.9}
+            tasks={tasks}
+            rowHeight={rowHeight}
+            svgWidth={svgWidth}
+            dates={dates}
+            todayColor={todayColor}
+            columnWidth={columnWidth}
+            rtl={rtl}
           />
         );
       }
@@ -302,6 +330,13 @@ export const Calendar: React.FC<CalendarProps> = ({
             y2Line={topDefaultHeight}
             xText={columnWidth * i + ticks * columnWidth * 0.5}
             yText={topDefaultHeight * 0.9}
+            tasks={tasks}
+            rowHeight={rowHeight}
+            svgWidth={svgWidth}
+            dates={dates}
+            todayColor={todayColor}
+            columnWidth={columnWidth}
+            rtl={rtl}
           />
         );
       }
@@ -349,6 +384,13 @@ export const Calendar: React.FC<CalendarProps> = ({
             y2Line={topDefaultHeight}
             xText={columnWidth * (i + topPosition)}
             yText={topDefaultHeight * 0.9}
+            tasks={tasks}
+            rowHeight={rowHeight}
+            svgWidth={svgWidth}
+            dates={dates}
+            todayColor={todayColor}
+            columnWidth={columnWidth}
+            rtl={rtl}
           />
         );
       }
@@ -357,92 +399,86 @@ export const Calendar: React.FC<CalendarProps> = ({
     return [topValues, bottomValues];
   };
 
-  const now = new Date();
-  let tickX = 0;
-  const ticks: ReactChild[] = [];
-  let today: ReactChild = <rect />;
-  for (let i = 0; i < dates.length; i++) {
-    const date = dates[i];
-    ticks.push(
-      <line
-        key={date.getTime()}
-        x1={tickX}
-        y1={0}
-        x2={tickX}
-        y2={y}
-        className={styles.gridTick}
-      />
-    );
-    if (
-      (i + 1 !== dates.length &&
-        date.getTime() < now.getTime() &&
-        dates[i + 1].getTime() >= now.getTime()) ||
-      // if current date is last
-      (i !== 0 &&
-        i + 1 === dates.length &&
-        date.getTime() < now.getTime() &&
-        addToDate(
-          date,
-          date.getTime() - dates[i - 1].getTime(),
-          "millisecond"
-        ).getTime() >= now.getTime())
-    ) {
-      today = (
-        <rect
-          x={tickX}
-          y={0}
-          width={columnWidth}
-          height={y}
-          fill={todayColor}
-        />
-      );
-    }
-    // rtl for today
-    if (
-      rtl &&
-      i + 1 !== dates.length &&
-      date.getTime() >= now.getTime() &&
-      dates[i + 1].getTime() < now.getTime()
-    ) {
-      today = (
-        <rect
-          x={tickX + columnWidth}
-          y={0}
-          width={columnWidth}
-          height={y}
-          fill={todayColor}
-        />
-      );
-    }
-    tickX += columnWidth;
-  }
+  // const now = new Date();
+  // let tickX = 0;
+  // const ticks: ReactChild[] = [];
+  // let today: ReactChild = <rect />;
+  // for (let i = 0; i < dates.length; i++) {
+  //   const date = dates[i];
+  //   ticks.push(
+  //     <line
+  //       key={date.getTime()}
+  //       x1={tickX}
+  //       y1={0}
+  //       x2={tickX}
+  //       y2={y}
+  //       className={styles.gridTick}
+  //     />
+  //   );
+  //   if (
+  //     (i + 1 !== dates.length &&
+  //       date.getTime() < now.getTime() &&
+  //       dates[i + 1].getTime() >= now.getTime()) ||
+  //     // if current date is last
+  //     (i !== 0 &&
+  //       i + 1 === dates.length &&
+  //       date.getTime() < now.getTime() &&
+  //       addToDate(
+  //         date,
+  //         date.getTime() - dates[i - 1].getTime(),
+  //         "millisecond"
+  //       ).getTime() >= now.getTime())
+  //   ) {
+  //     today = (
+  //       <rect
+  //         x={tickX}
+  //         y={0}
+  //         width={columnWidth}
+  //         height={y}
+  //         fill={todayColor}
+  //       />
+  //     );
+  //   }
+  //   // rtl for today
+  //   if (
+  //     rtl &&
+  //     i + 1 !== dates.length &&
+  //     date.getTime() >= now.getTime() &&
+  //     dates[i + 1].getTime() < now.getTime()
+  //   ) {
+  //     today = (
+  //       <rect
+  //         x={tickX + columnWidth}
+  //         y={0}
+  //         width={columnWidth}
+  //         height={y}
+  //         fill={todayColor}
+  //       />
+  //     );
+  //   }
+  //   tickX += columnWidth;
+  // }
 
   let topValues: ReactChild[] = [];
   let bottomValues: ReactChild[] = [];
   switch (dateSetup.viewMode) {
     case ViewMode.Year:
-      <g className="ticks">{ticks}</g>
       [topValues, bottomValues] = getCalendarValuesForYear();
       break;
       case ViewMode.Month:
-        <g className="ticks">{ticks}</g>
         [topValues, bottomValues] = getCalendarValuesForMonth();
         break;
       case ViewMode.Week:
-      <g className="ticks">{ticks}</g>
       [topValues, bottomValues] = getCalendarValuesForWeek();
       break;
     case ViewMode.Day:
-      <g className="ticks">{ticks}</g>
       [topValues, bottomValues] = getCalendarValuesForDay();
       break;
     case ViewMode.QuarterDay:
     case ViewMode.HalfDay:
-      <g className="ticks">{ticks}</g>
       [topValues, bottomValues] = getCalendarValuesForPartOfDay();
       break;
     case ViewMode.Hour:
-      <g className="ticks">{ticks}</g>
       [topValues, bottomValues] = getCalendarValuesForHour();
   }
   return (
@@ -462,7 +498,6 @@ export const Calendar: React.FC<CalendarProps> = ({
         className={styles.calendarHeader}
       />
       {bottomValues} {topValues}
-      <g className="today">{today}</g>
     </g>
   );
 };
